rename file, function and variable
