+++ api/v1/__init__.py
@@ -9,7 +9,7 @@ from .register import register as router_register
 from .authentication import authentication as router_authentication
 from .schedule import schedule as router_schedule
 from .role import role as router_role
-from .information_user import view as router_view
+from .view_information import view as router_view
 
 from fastapi import APIRouter
