+++ api/v1/register.py
@@ -17,7 +17,7 @@ async def register_user(
     data: UserCreateSchema,
     db: AsyncSession = Depends(get_async_db_session),
 ):
-    services = RegisterService(db)
+    register_service = RegisterService(db)
 
     if data.role_id not in ROLE_MAPPING_REGISTER_SCHEMA:
         raise HTTPException(status_code=400, detail="Invalid role")
@@ -25,7 +25,7 @@ async def register_user(
     schema_class, method_name = ROLE_MAPPING_REGISTER_SCHEMA[data.role_id]
     user = schema_class.model_validate(data)
 
-    service_method = getattr(services, method_name)
+    service_method = getattr(register_service, method_name)
     if not callable(service_method):
         raise HTTPException(
             status_code=500, detail=f"Service method '{method_name}' not found"
